requirements.txt
pymongo==4.6.1
python-dateutil==2.8.2
colorama==0.4.6
tabulate==0.9.0 